Изменение логики при добавлении/изменении пользователя (роль Юзер ставится по дефолту). Добавлен RestExceptionHandler для обработки ошибок. Доработаны методы RestController, Добавлены, но пока закомментированы настройки доступа к рест контроллеру в SecurityConfig

+++ src/main/java/ru.atavrel.springbootapp/rest/UserRestController.java
@@ -1,7 +1,9 @@
 package ru.atavrel.springbootapp.rest;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import ru.atavrel.springbootapp.exception.ResourceNotFoundException;
 import ru.atavrel.springbootapp.model.User;
 import ru.atavrel.springbootapp.service.UserService;
 
@@ -19,10 +21,7 @@ class UserRestController {
     }
 
     //todo: прописать настройки доступа к api в security
-    //todo: добавить или нет обработку ошибок при работе с неверными данными?
-    // + возвращать правильные статусы?
-    // + для update передавать Id или нет ?
-
+    // + разобраться детально что делает мой класс RestResponseEntityExceptionHandler
 
     // GET: http://localhost:8080/api/users
     @GetMapping
@@ -33,26 +32,42 @@ class UserRestController {
     //GET: http://localhost:8080/api/users/id
     @GetMapping("{id}")
     public User getById(@PathVariable Long id) {
-        return userService.getById(id);
+        User user = userService.getById(id);
+        if (user == null) {
+            throw new ResourceNotFoundException("No user found with such id: " + id);
+        }
+        return user;
     }
 
     //POST: http://localhost:8080/api/users/
     @PostMapping
-    public boolean save(@RequestBody User user) {
-        return userService.add(user);
+    @ResponseStatus(HttpStatus.CREATED)
+    public User save(@RequestBody User user) {
+        userService.add(user);
+        return user;
     }
 
     //PUT: http://localhost:8080/api/users/id
     @PutMapping("{id}")
-    public void updateById(@PathVariable("id") Long id, @RequestBody User user) {
-        // что делать с id ? (пока просто устанавливаю id юзеру)
+    @ResponseStatus(HttpStatus.OK)
+    public User updateById(@PathVariable("id") Long id, @RequestBody User user) {
+        User userFromDB = userService.getById(id);
+        if (userFromDB == null) {
+            throw new ResourceNotFoundException("No user found with such id: " + id);
+        }
         user.setId(id);
         userService.update(user);
+        return user;
     }
 
     //DELETE: http://localhost:8080/api/users/id
     @DeleteMapping("{id}")
+    @ResponseStatus(HttpStatus.OK)
     public void deleteById(@PathVariable Long id) {
+        User user = userService.getById(id);
+        if (user == null) {
+            throw new ResourceNotFoundException("No user found with such id: " + id);
+        }
         userService.deleteById(id);
     }
